[webui] improve webui & reasoning mode (#6778)
