+++ src/llamafactory/webui/locales.py
@@ -2786,6 +2786,20 @@ ALERTS = {
         "ko": "모델이 언로드되었습니다.",
         "ja": "モデルがアンロードされました。",
     },
+    "info_thinking": {
+        "en": "🌀 Thinking...",
+        "ru": "🌀 Думаю...",
+        "zh": "🌀 思考中...",
+        "ko": "🌀 생각 중...",
+        "ja": "🌀 考えています...",
+    },
+    "info_thought": {
+        "en": "✅ Thought",
+        "ru": "✅ Думать закончено",
+        "zh": "✅ 思考完成",
+        "ko": "✅ 생각이 완료되었습니다",
+        "ja": "✅ 思考完了",
+    },
     "info_exporting": {
         "en": "Exporting model...",
         "ru": "Экспорт модели...",
